#137 Discrete feature map compiles. Bag feature map and discrete removed because I can see no elegant way of including discreteness. Continuous control problems is our goal, discrete ones are only a bonus

+++ RLSimion/controller.h
@@ -14,8 +14,8 @@ class Controller : public Simion
 
 public:
 	virtual ~Controller(){}
-	virtual int getNumOutputs()= 0;
-	virtual int getOutputActionIndex(int output)= 0;
+	virtual size_t getNumOutputs()= 0;
+	virtual size_t getOutputActionIndex(int output)= 0;
 
 	static std::shared_ptr<Controller> getInstance(ConfigNode* pConfigNode);
 
@@ -43,8 +43,8 @@ public:
 	LQRController(ConfigNode* pConfigNode);
 	virtual ~LQRController();
 
-	int getNumOutputs();
-	int getOutputActionIndex(int output);
+	size_t getNumOutputs();
+	size_t getOutputActionIndex(int output);
 
 	double selectAction(const State *s,Action *a);
 	virtual double update(const State *s, const Action *a, const State *s_p, double r, double behaviorProb) { return 1.0; }
@@ -62,8 +62,8 @@ public:
 	PIDController(ConfigNode* pConfigNode);
 	virtual ~PIDController();
 
-	int getNumOutputs();
-	int getOutputActionIndex(int output);
+	size_t getNumOutputs();
+	size_t getOutputActionIndex(int output);
 
 	double selectAction(const State *s,Action *a);
 	virtual double update(const State *s, const Action *a, const State *s_p, double r, double behaviorProb) { return 1.0; }
@@ -76,12 +76,12 @@ protected:
 	double sgn(double value);
 
 	//state variable indices
-	int m_omega_g, m_d_omega_g;
-	int m_omega_r, m_d_omega_r;
-	int m_E_p, m_T_g, m_beta;
-	int m_E_int_omega_r;
+	size_t m_omega_g, m_d_omega_g;
+	size_t m_omega_r, m_d_omega_r;
+	size_t m_E_p, m_T_g, m_beta;
+	size_t m_E_int_omega_r;
 	//action variable indices
-	int m_a_beta, m_a_T_g;
+	size_t m_a_beta, m_a_T_g;
 	double m_ratedPower;
 	double m_genElecEff;
 	double m_lastT_g = 0.0;
@@ -90,8 +90,8 @@ public:
 	WindTurbineVidalController(ConfigNode* pConfigNode);
 	virtual ~WindTurbineVidalController();
 
-	int getNumOutputs();
-	int getOutputActionIndex(int output);
+	size_t getNumOutputs();
+	size_t getOutputActionIndex(int output);
 
 	virtual double selectAction(const State *s,Action *a);
 	virtual double update(const State *s, const Action *a, const State *s_p, double r, double behaviorProb) { return 1.0; }
@@ -101,12 +101,12 @@ class WindTurbineBoukhezzarController : public Controller
 {
 protected:
 	//state variable indices
-	int m_omega_g, m_d_omega_g;
-	int m_E_p, m_T_a_index, m_T_g, m_d_T_g;
-	int m_beta, m_E_int_omega_g;
+	size_t m_omega_g, m_d_omega_g;
+	size_t m_E_p, m_T_a_index, m_T_g, m_d_T_g;
+	size_t m_beta, m_E_int_omega_g;
 
 	//action variable indices
-	int m_a_beta, m_a_T_g;
+	size_t m_a_beta, m_a_T_g;
 	
 	DOUBLE_PARAM m_pC_0, m_pKP, m_pKI;
 	double m_K_t, m_J_t;
@@ -116,8 +116,8 @@ public:
 	WindTurbineBoukhezzarController(ConfigNode* pConfigNode);
 	virtual ~WindTurbineBoukhezzarController();
 
-	int getNumOutputs();
-	int getOutputActionIndex(int output);
+	size_t getNumOutputs();
+	size_t getOutputActionIndex(int output);
 
 	virtual double selectAction(const State *s,Action *a);
 	virtual double update(const State *s, const Action *a, const State *s_p, double r, double behaviorProb) { return 1.0; }
@@ -126,9 +126,9 @@ public:
 class WindTurbineJonkmanController : public Controller
 {
 protected:
-	int m_omega_g, m_d_omega_g;
-	int m_E_p, m_T_g, m_beta;
-	int m_a_beta, m_a_T_g;
+	size_t m_omega_g, m_d_omega_g;
+	size_t m_E_p, m_T_g, m_beta;
+	size_t m_a_beta, m_a_T_g;
 
 	//generator speed filter's parameters and variables
 	DOUBLE_PARAM m_CornerFreq;
@@ -146,8 +146,8 @@ public:
 	WindTurbineJonkmanController(ConfigNode* pConfigNode);
 	virtual ~WindTurbineJonkmanController();
 
-	int getNumOutputs();
-	int getOutputActionIndex(int output);
+	size_t getNumOutputs();
+	size_t getOutputActionIndex(int output);
 
 	virtual double selectAction(const State *s,Action *a);
 	virtual double update(const State *s, const Action *a, const State *s_p, double r, double behaviorProb) { return 1.0; }
